requirements.txt
Scrapy
Pymongo
mysql-connector-python